test: Move modinv to util and add unit test
